% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomatch.R
\name{geo_match}
\alias{geo_match}
\title{Match Geographies Based on Statistical Similarity}
\usage{
geo_match(
  .data,
  id_col,
  target,
  covariates,
  .matches = 1,
  .scale = FALSE,
  .keep = FALSE
)
}
\arguments{
\item{.data}{A \code{data.frame} or \link[tibble:tibble-package]{tibble} with one
character column of unique geographic labels and at least one numeric
column of a feature about those geographies.}

\item{id_col}{The (unquoted) column name of geographic labels.}

\item{target}{The target geographic label to which a match will be made.}

\item{covariates}{The (quoted) column, or columns (in the form \code{c(...)})
of features to use for calculating the similarity metric.}

\item{.matches}{(Optional) The number of geographies matched to \code{target} to
return. Defaults to 1.}

\item{.scale}{(Optional) Whether to normalize the \code{covariates} before
calculating the similarity metric. Defaults to \code{FALSE}.}

\item{.keep}{(Optional) Whether to keep all columns in \code{.data} in the output
(\code{TRUE}) or just the \code{id_col} and \code{covariates} (\code{FALSE}). Defaults to
\code{FALSE}.}
}
\value{
A \code{.matches} + 1 row \link[tibble:tibble-package]{tibble} of \code{id_col},
\code{covariates}, and an additional \code{.distance} column that contains the
calculated similarity metric.
}
\description{
This function takes a \code{data.frame} of geographies (\emph{e.g.}, U.S. States, ZIP
codes, \emph{etc.}), where each row is a unique geography, and features about
those geographies (\emph{e.g.}, population density, GDP, \emph{etc.}), and, for one
specified geography, calculates the "most similar" geography to it. That is,
given a \code{data.frame} of U.S. States, and various features about those States,
and specifying one specific state (say, Massachusetts), this function will
calculate a similarity metric (\emph{e.g.}, Euclidean Distance) and return, from
the remaining States, the State most similar to it.
}
\details{
\subsection{Feature Format Requirements}{

Note that, in order to calculate the similarity metric, all features must be
numeric. Non-numeric features will be ignored. Note also that the column
of geographic labels must be character. If a non-character column is
supplied, it will be converted to character.
}

\subsection{Available Similarity Metrics}{

Currently, Euclidean Distance is used as the similarity metric. Future
updates will include additional metrics to choose.
}
}
\examples{
\dontrun{
geo_match(simulated_states, state, "CT", c("avg_age", "pop_density"))

geo_match(
  simulated_states, state, "CT", c("avg_age", "pop_density"), .matches = 3
)

geo_match(
  simulated_states, state, "CT", c("avg_age", "pop_density"), .scale = 3
)
}

}
