% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomatch-package.R
\docType{package}
\name{geomatch-package}
\alias{geomatch}
\alias{geomatch-package}
\title{geomatch: Geography-Focused Matching Methods}
\description{
Provides functions for matching geographies (state, zip code, etc.) based on geographic-specific features.
}
\author{
\strong{Maintainer}: Robert Lankford \email{rob.j.lankford@gmail.com}

}
\keyword{internal}
