% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMatch.R
\docType{package}
\name{geoMatch}
\alias{geoMatch}
\alias{geoMatch-package}
\title{geoMatch}
\description{
A package designed to provide matched data for propensity-score based causal analyses using spatially-explicit data
}

