% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_geographies}
\alias{simulated_geographies}
\alias{simulated_zips}
\alias{simulated_states}
\title{Simulated Data by Geography}
\format{
\subsection{\code{simulated_zips}}{

A \code{tibble} with 30 rows and 8 columns:
\describe{
\item{zip}{Five-digit ZIP code (possibly not a real ZIP code)}
\item{med_income}{Median income}
\item{pop_density}{Population density}
\item{pct_college}{Percent of population with a college degree}
\item{avg_age}{Average age of population}
\item{pct_bcollar}{Percent of population with blue collar job}
\item{prior_spend}{Amount of marketing spend in prior quarter}
\item{brand_aware}{Percent of population aware of the brand}
}
}

\subsection{\code{simulated_states}}{

A \code{tibble} with 30 rows and 8 columns (7 are the same as \code{simulated_zips}):
\describe{
\item{state}{Two-digit state abbreviation}
}
}
}
\usage{
simulated_zips

simulated_states
}
\description{
These data sets provide simulated geographic level features. They are toy
data sets used for internal testing, examples, and demonstrations. Note
that this data is simulated and, therefore, \strong{not real}.
}
\details{
\code{simulated_zips} is simulated at a ZIP code level and \code{simulated_states} is
simulated at a state level.
}
\keyword{datasets}
